Integrate CUSUM methods into change detection
