+++ frouros/detectors/concept_drift/streaming/change_detection/__init__.py
@@ -1,8 +1,29 @@
 """Concept drift change detection methods' init."""
 
-from .bocd import BOCD, BOCDConfig
+from .bocd import (
+    BOCD,
+    BOCDConfig,
+)
+from .cusum import (
+    CUSUM,
+    CUSUMConfig,
+)
+from .geometric_moving_average import (
+    GeometricMovingAverage,
+    GeometricMovingAverageConfig,
+)
+from .page_hinkley import (
+    PageHinkley,
+    PageHinkleyConfig,
+)
 
 __all__ = [
     "BOCD",
     "BOCDConfig",
+    "CUSUM",
+    "CUSUMConfig",
+    "GeometricMovingAverage",
+    "GeometricMovingAverageConfig",
+    "PageHinkley",
+    "PageHinkleyConfig",
 ]
