+++ frouros/detectors/concept_drift/streaming/change_detection/page_hinkley.py
@@ -1,16 +1,15 @@
-"""CUSUM module."""
+"""Page Hinkley module."""
 
-import numpy as np  # type: ignore
-
-from frouros.detectors.concept_drift.streaming.cusum_based.base import (
+from frouros.detectors.concept_drift.streaming.change_detection.base import (
     BaseCUSUM,
     BaseCUSUMConfig,
     DeltaConfig,
+    AlphaConfig,
 )
 
 
-class CUSUMConfig(BaseCUSUMConfig, DeltaConfig):
-    """CUSUM [page1954continuous]_ configuration.
+class PageHinkleyConfig(BaseCUSUMConfig, DeltaConfig, AlphaConfig):
+    """Page Hinkley [page1954continuous]_ configuration.
 
     :References:
 
@@ -24,25 +23,29 @@ class CUSUMConfig(BaseCUSUMConfig, DeltaConfig):
         delta: float = 0.005,
         lambda_: float = 50.0,
         min_num_instances: int = 30,
+        alpha: float = 0.9999,
     ) -> None:
         """Init method.
 
         :param delta: delta value
         :type delta: float
-        :param lambda_: delta value
+        :param lambda_: lambda value
         :type lambda_: float
         :param min_num_instances: minimum numbers of instances
         to start looking for changes
         :type min_num_instances: int
+        :param alpha: forgetting factor value
+        :type alpha: float
         """
         BaseCUSUMConfig.__init__(
-            self, lambda_=lambda_, min_num_instances=min_num_instances
+            self, min_num_instances=min_num_instances, lambda_=lambda_
         )
         DeltaConfig.__init__(self, delta=delta)
+        AlphaConfig.__init__(self, alpha=alpha)
 
 
-class CUSUM(BaseCUSUM):
-    """CUSUM [page1954continuous]_ detector.
+class PageHinkley(BaseCUSUM):
+    """Page Hinkley [page1954continuous]_ detector.
 
     :References:
 
@@ -51,13 +54,9 @@ class CUSUM(BaseCUSUM):
         Biometrika 41.1/2 (1954): 100-115.
     """
 
-    config_type = CUSUMConfig  # type: ignore
+    config_type = PageHinkleyConfig  # type: ignore
 
     def _update_sum(self, error_rate: float) -> None:
-        self.sum_ = np.maximum(
-            0,
-            self.sum_
-            + error_rate
-            - self.mean_error_rate.mean
-            - self.config.delta,  # type: ignore
+        self.sum_ = self.config.alpha * self.sum_ + (  # type: ignore
+            error_rate - self.mean_error_rate.mean - self.config.delta  # type: ignore
         )
