--- frouros/detectors/concept_drift/streaming/cusum_based/__init__.py
@@ -1,17 +1,0 @@
-"""Concept drift CUSUM based detection methods' init."""
-
-from .cusum import CUSUM, CUSUMConfig
-from .geometric_moving_average import (
-    GeometricMovingAverage,
-    GeometricMovingAverageConfig,
-)
-from .page_hinkley import PageHinkley, PageHinkleyConfig
-
-__all__ = [
-    "CUSUM",
-    "CUSUMConfig",
-    "GeometricMovingAverage",
-    "GeometricMovingAverageConfig",
-    "PageHinkley",
-    "PageHinkleyConfig",
-]
